+++ commons/commons-model-extended-serialization/src/main/kotlin/uk/ac/ebi/extended/serialization/service/ExtSerializationService.kt
@@ -58,7 +58,7 @@ class ExtSerializationService {
     /**
      * Serialize a generic element. ONLY for testing purpose.
      */
-    internal fun <T> serializeElement(element: T, properties: Properties = Properties(false)): String {
+    private internal fun <T> serializeElement(element: T, properties: Properties = Properties(false)): String {
         mapper.writerWithDefaultPrettyPrinter().writeValue(properties, element)
         return properties.buffer.toString()
     }
