+++ commons/commons-serialization/src/main/kotlin/ac/uk/ebi/biostd/tsv/TsvSerializer.kt
@@ -3,13 +3,13 @@ package ac.uk.ebi.biostd.tsv
 import ac.uk.ebi.biostd.tsv.deserialization.TsvDeserializer
 import ac.uk.ebi.biostd.tsv.deserialization.stream.FileListTsvStreamDeserializer
 import ac.uk.ebi.biostd.tsv.deserialization.stream.PageTabFile
-import ac.uk.ebi.biostd.tsv.serialization.TsvToStringSerializer
+import ac.uk.ebi.biostd.tsv.serialization.TsvSerializer
 import ebi.ac.uk.model.Submission
 import java.io.InputStream
 import java.io.OutputStream
 
 internal class TsvSerializer(
-    private val tsvSerializer: TsvToStringSerializer = TsvToStringSerializer(),
+    private val tsvSerializer: TsvSerializer = TsvSerializer(),
     private val tsvDeserializer: TsvDeserializer = TsvDeserializer(),
     private val streamSerializer: FileListTsvStreamDeserializer = FileListTsvStreamDeserializer()
 ) {
