+++ commons/commons-serialization/src/test/kotlin/ac/uk/ebi/biostd/test/TsvTestFactory.kt
@@ -105,8 +105,8 @@ fun submissionWithMultipleLineBreaks() = tsv {
 
 fun submissionWithSectionsTable() = submissionWithRootSection().apply {
     line("Data[]", "Title", "Desc")
-    line("DT-1", "Data 1", "Group 1")
-    line("DT-2", "Data 2", "Group 2")
+    line("DT-1", "", "Group 1")
+    line("DT-2", "Data 2", "")
     line()
 }
